Failing case of Verilated UART 16550
Added examples of compiling a Verilated UART using the current UART
VIL and the proposed new one.  The Robot file in this commit invokes
the scripts to call the failing case in smm_mmuart_uart/Vtop_uart

Renode portable is intended to be run from the base directory in this
repo.

Signed-off-by: Edward Connell <Edward.Connell@microchip.com>